remove rewrite on local entrypoint
